Refactor prompts so each prompt is isolated
Pair the prompt with its schema in a single file with its parsing. Isolates individual prompts for easier updating.
